// src/core/services/api/httpType.ts
import { IContextSession } from '../../domain/session/sessionTypes';
import { IError } from '../../domain/common/errorType';

export enum HttpStatusCode {
    OK = 200,
    CREATED = 201,
    ACCEPTED = 202,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    CONFLICT = 409,
    INTERNAL_SERVER_ERROR = 500
}

export interface BApiError extends IError {

    session: IContextSession;
    httpStatusCode: HttpStatusCode;

}